+++ sn_client/src/lib.rs
@@ -15,7 +15,7 @@ mod chunks;
 mod error;
 mod event;
 mod faucet;
-mod file_apis;
+mod files;
 mod register;
 mod wallet;
 
@@ -25,7 +25,7 @@ pub use self::{
     error::Error,
     event::{ClientEvent, ClientEventsReceiver},
     faucet::{get_tokens_from_faucet, load_faucet_wallet_from_genesis_wallet},
-    file_apis::{Files, BATCH_SIZE},
+    files::{api::FilesApi, FileUploadEvent, Files, BATCH_SIZE, MAX_UPLOAD_RETRIES},
     register::ClientRegister,
     wallet::{send, WalletClient},
 };