+++ sn_client/CHANGELOG.md
@@ -6,6 +6,39 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.99.25](https://github.com/maidsafe/safe_network/compare/sn_client-v0.99.24...sn_client-v0.99.25) - 2023-12-19
+
+### Fixed
+- *(test)* tests should try to load just the faucet wallet
+
+## [0.99.24](https://github.com/maidsafe/safe_network/compare/sn_client-v0.99.23...sn_client-v0.99.24) - 2023-12-19
+
+### Other
+- updated the following local packages: sn_networking
+
+## [0.99.23](https://github.com/maidsafe/safe_network/compare/sn_client-v0.99.22...sn_client-v0.99.23) - 2023-12-19
+
+### Fixed
+- *(cli)* mark chunk completion as soon as we upload each chunk
+
+## [0.99.22](https://github.com/maidsafe/safe_network/compare/sn_client-v0.99.21...sn_client-v0.99.22) - 2023-12-18
+
+### Added
+- *(transfers)* add api for cleaning up CashNotes
+
+## [0.99.21](https://github.com/maidsafe/safe_network/compare/sn_client-v0.99.20...sn_client-v0.99.21) - 2023-12-18
+
+### Added
+- *(client)* update the Files config via setters
+- *(client)* track the upload stats inside Files
+- *(client)* move upload retry logic from CLI to client
+
+### Fixed
+- *(test)* use the Files struct to upload chunks
+
+### Other
+- *(client)* add docs to the Files struct
+
 ## [0.99.20](https://github.com/maidsafe/safe_network/compare/sn_client-v0.99.19...sn_client-v0.99.20) - 2023-12-14
 
 ### Other