+++ sn_faucet/CHANGELOG.md
@@ -6,6 +6,41 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.96](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.95...sn_faucet-v0.1.96) - 2023-12-19
+
+### Other
+- update dependencies
+
+## [0.1.95](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.94...sn_faucet-v0.1.95) - 2023-12-19
+
+### Other
+- update dependencies
+
+## [0.1.94](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.93...sn_faucet-v0.1.94) - 2023-12-19
+
+### Other
+- update dependencies
+
+## [0.1.93](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.92...sn_faucet-v0.1.93) - 2023-12-18
+
+### Other
+- update dependencies
+
+## [0.1.92](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.91...sn_faucet-v0.1.92) - 2023-12-18
+
+### Other
+- update dependencies
+
+## [0.1.91](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.90...sn_faucet-v0.1.91) - 2023-12-18
+
+### Other
+- update dependencies
+
+## [0.1.90](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.89...sn_faucet-v0.1.90) - 2023-12-18
+
+### Other
+- update dependencies
+
 ## [0.1.89](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.88...sn_faucet-v0.1.89) - 2023-12-14
 
 ### Other